[Update] Fixed arne's murge

+++ src/GUILogic/SimulatorLogic/NPCLogic/Person.java
@@ -36,13 +36,13 @@ public class Person {
      * @param genreChanceList a list for the probability of liking a Genre
      * @param speed           the movement speed of the NPCLogic.Person
      */
-    public Person(Point2D position, ArrayList<Integer> genreChanceList, int speed, boolean isArtist) {
+    public Person(Point2D position, ArrayList<Integer> genreChanceList, double speed, boolean isArtist) {
         this.isArtist = isArtist;
         genrePicker(genreChanceList);
         this.personLogic = new PersonLogic(position, speed, this, isArtist);
     }
 
-    public Person(Point2D position, ArrayList<Integer> genreChanceList, String name, int speed, boolean isArtist) {
+    public Person(Point2D position, ArrayList<Integer> genreChanceList, String name, double speed, boolean isArtist) {
         this.name = name;
         this.isArtist = isArtist;
         genrePicker(genreChanceList);
